requirements.txt
flask
flask-cors
python-dotenv
pinecone-client
langchain
langchain-openai
langchain-pinecone
openai